% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{read.tzh-package}
\alias{read.tzh}
\alias{read.tzh-package}
\title{read.tzh}
\description{
..
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} [copyright holder]

}
\keyword{internal}
