% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_ext.R
\name{read_csv_txt}
\alias{read_csv_txt}
\alias{read_csv}
\alias{read_txt}
\title{Read Big \code{.csv} and \code{.txt} File}
\usage{
read_csv(file, header = TRUE, ...)

read_txt(file, header = TRUE, ...)
}
\arguments{
\item{file}{\link[base]{character} scalar, file name (extension not needed).}

\item{header}{\link[base]{logical} scalar}

\item{...}{..}
}
\description{
..
}
\note{
\link[data.table]{fread} returns \link[data.table]{data.table} object
}
