% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxl_ext.R
\name{read_excel_ext}
\alias{read_excel_ext}
\alias{read_excel_pattern}
\alias{read_excel_sheets}
\alias{read_excel_all}
\title{Read Local Excel File}
\usage{
read_excel_pattern(path = stop(), pattern, ...)

read_excel_sheets(path = stop(), sheets, ...)

read_excel_all(path = stop(), ...)
}
\arguments{
\item{path, ...}{parameters of \link[readxl]{read_excel}}

\item{pattern}{regular expression, pattern of the names of sheets to be selected}

\item{sheets}{\link[base]{integer} \link[base]{vector}}
}
\description{
..
}
\details{
Function \link{read_excel_all} ..

Function \link{read_excel_sheets} ..

Function \link{read_excel_pattern} ..
}
\note{
\link[readxl]{read_excel} returns (a \link[base]{list} of) \link[tibble]{tibble} object(s).

Name clash \link[officer]{read_xlsx} and \link[readxl]{read_xlsx}.

\CRANpkg{readxl} cannot write to Excel file.  Write to \code{.csv} file instead.
}
\references{
\url{https://cran.r-project.org/web/packages/Microsoft365R/vignettes/od_sp.html}
can access data stored in SharePoint Online sites and OneDrive (personal and business).
But need to request approval from Jefferson.
}
